image-localization-system/backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Image Processing
Pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.3

# OCR
pytesseract==0.3.10
easyocr==1.7.1

# AI/ML
anthropic==0.18.1
openai==1.12.0
transformers==4.37.2
torch==2.2.0
sentence-transformers==2.3.1

# Background Removal
rembg==2.0.56

# Task Queue
celery==5.3.6
redis==5.0.1

# Storage
boto3==1.34.34
google-cloud-storage==2.14.0

# Utilities
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0
tenacity==8.2.3

# Monitoring
sentry-sdk==1.40.0
prometheus-client==0.19.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
